requirements.txt
flask
garminconnect>=0.3.2
curl_cffi
python-dotenv
gunicorn
google-genai
tzdata
